+++ packages/core/src/constants/request-origins.js
@@ -15,7 +15,7 @@ const forServerSideRendering = {
   },
   [releaseBranch.staging]: {
     accounts: 'https://staging-accounts.twreporter.org',
-    api: 'http://go-api:8080',
+    api: 'https://staging-go-api.twreporter.org',
     main: 'http://staging.twreporter.org',
     support: 'https://staging-support.twreporter.org',
   },
@@ -27,45 +27,25 @@ const forServerSideRendering = {
   },
   [releaseBranch.release]: {
     accounts: 'https://accounts.twreporter.org',
-    api: 'http://go-api:8080',
+    api: 'https://go-api.twreporter.org',
     main: 'https://www.twreporter.org',
     support: 'https://support.twreporter.org',
   },
-  [releaseBranch.next]: {
-    accounts: 'https://next-accounts.twreporter.org',
-    api: 'http://go-api:8080',
-    main: 'https://next.twreporter.org',
-    support: 'https://next-support.twreporter.org',
+  [releaseBranch.dev]: {
+    accounts: 'https://dev-accounts.twreporter.org',
+    api: 'https://staging-go-api.twreporter.org',
+    main: 'https://dev.twreporter.org',
+    support: 'https://dev-support.twreporter.org',
   },
 }
 
 const forClientSideRendering = {
   [releaseBranch.master]: forServerSideRendering[releaseBranch.master],
   [releaseBranch.test]: forServerSideRendering[releaseBranch.test],
-  [releaseBranch.staging]: {
-    accounts: forServerSideRendering[releaseBranch.staging].accounts,
-    api: 'https://staging-go-api.twreporter.org',
-    main: forServerSideRendering[releaseBranch.staging].main,
-    support: forServerSideRendering[releaseBranch.staging].support,
-  },
-  [releaseBranch.preview]: {
-    accounts: forServerSideRendering[releaseBranch.preview].accounts,
-    api: 'https://go-api.twreporter.org',
-    main: forServerSideRendering[releaseBranch.preview].main,
-    support: forServerSideRendering[releaseBranch.preview].support,
-  },
-  [releaseBranch.release]: {
-    accounts: forServerSideRendering[releaseBranch.release].accounts,
-    api: 'https://go-api.twreporter.org',
-    main: forServerSideRendering[releaseBranch.release].main,
-    support: forServerSideRendering[releaseBranch.release].support,
-  },
-  [releaseBranch.next]: {
-    accounts: forServerSideRendering[releaseBranch.next].accounts,
-    api: 'https://next-go-api.twreporter.org',
-    main: forServerSideRendering[releaseBranch.next].main,
-    support: forServerSideRendering[releaseBranch.next].support,
-  },
+  [releaseBranch.staging]: forServerSideRendering[releaseBranch.staging],
+  [releaseBranch.preview]: forServerSideRendering[releaseBranch.preview],
+  [releaseBranch.release]: forServerSideRendering[releaseBranch.release],
+  [releaseBranch.dev]: forServerSideRendering[releaseBranch.dev],
 }
 
 export default {