+++ packages/react-components/src/input/enums/index.js
@@ -1,4 +1,17 @@
+// enums for search-bar
 export const WidthType = Object.freeze({
   FIT: 'fit',
   STRETCH: 'stretch',
+})
+
+// enums for text-field
+export const TextAlign = Object.freeze({
+  LEFT: 'left',
+  CENTER: 'center',
+})
+
+export const TextState = Object.freeze({
+  DEFAULT: 'default',
+  ERROR: 'error',
+  DISABLED: 'disabled',
 })